Support virtual modules
This `dune` feature allows interfaces to be declared as `virtual`,
meaning that they have no corresponding module in their library.
Modules that implement this virtual interface can be defined in separate
libraries and are only expected to be linked into executables.

Adds a new attr `virtual` to `ocaml_signature`, which is analyzed by
module rules to adapt the module name prefix (for NS/wrapped libraries)
to be equal to the interface, since `ocamlc` requires this to associate
them with each other.